+++ src/image.rs
@@ -5,21 +5,21 @@ use crate::polygon;
 
 /// Star position in image
 pub struct ImageStar {
-    pub pixel_x: f32,
-    pub pixel_y: f32,
-    pub magnitude: f32
+    pub pixel_x: f64,
+    pub pixel_y: f64,
+    pub magnitude: f64
 }
 
 /// Image metadata
 pub struct Image {
     pub height: u32,
     pub width: u32,
-    pub scale_ppa: f32, // Pixels per arcsecond
+    pub scale_ppa: f64, // Pixels per arcsecond
     pub star_list: Vec<ImageStar>
 }
 
 /// Converts the image::Star structure to polygon::Star
-pub fn image_star_to_polygon(star_list: &Vec<ImageStar>, scale_app: f32) -> Vec<polygon::Star> {
+pub fn image_star_to_polygon(star_list: &Vec<ImageStar>, scale_app: f64) -> Vec<polygon::Star> {
     let scale_rad = math::degrees_to_radians(scale_app / 60.0 / 60.0);
     let mut pol_star_list = Vec::new();
     debug!("Image Star to Polygon > Star list:{} Scale \"pp:{} Scale rpp:{}", star_list.len(), scale_app, scale_rad);
